compile under g++ 3.4.2
git-svn-id: http://www.lunabase.org/svn/grap/trunk@282 17fcd644-1d9d-df11-a57b-001cc0feadb3

+++ grap_tokenizer.cc
@@ -286,7 +286,7 @@ void init_id() {
     for ( i = 'a'; i <= 'z' ; i++ ) id_letter[i] = true;
     for ( i = 'A'; i <= 'Z' ; i++ ) id_letter[i] = true;
     for ( i = '0'; i <= '9' ; i++ ) id_letter[i] = true;
-    id_letter['_'] = true;
+    id_letter[static_cast<unsigned int>('_')] = true;
 }
 
 // Initialize the table of keywords.
